bound board dimensins to model, added model for main viewmodel, added anglevector, added vectorball

+++ Data/Position/IPosition.cs
@@ -6,8 +6,8 @@ namespace Data.Position
 {
     public interface IPosition
     {
-        float X { get; }
+        double X { get; }
 
-        float Y { get; }
+        double Y { get; }
     }
 }
